Switch contact-us-pages-dynamic mapper to ts
